HMM training works finally.
I have found one of the source of the error. Its the zeros-sample in the semitones that cause NaN value in the forward proceding. So I have add 1 to the samples of the semitones.
*******
Problem 2 (corrected) problem for the dimensions of c in MarkovChain/AdaptAccum. I put a tranpose to fix it: it works